services/blog_engine/app/services/image/utils.py
from PIL import ImageFont
from .image_config import *

def load_font(size: int):
    try:
        return ImageFont.truetype(str(FONT_PATH), size)
    except Exception:
        return ImageFont.load_default()
