services/blog_engine/requirements.txt
jinja2
pillow
beautifulsoup4
lxml
requests
numpy